refactor: move deriveKey to crypto utils
